ramp-up/cool-down
Implement logic to adjust the behavior depending on how much
temperature still rises after turning the heat off, or how
much it lowers after turning the heat on.

It supports an initial value, but it will continue adjusting
as it goes so it can better fit the goal variance.

This has managed to very precisely keep the temperature within
the configured variance band.